Fixing pubsub monitor

+++ flask_monitor/db_monitor.py
@@ -161,7 +161,7 @@ def db_check_redis_pubsub():
     except:
         abort(504)
 
-    if pubsub_numpat or isinstance(pubsub_numpat, long):
+    if pubsub_numpat or isinstance(pubsub_numpat, int):
         return str(pubsub_numpat)
     else:
         abort(505)
